+++ MOVIDIUS/Test/inputsize.txt
@@ -1,0 +1,1 @@
+224
