+++ MOVIDIUS/Test/movidiustest.py
